Now using threads
Update the initial tool to use threads. There is currently a ReadThread
which reads the input files and publishes their value to a global
variable. A convert thread which reads from that global variable and
prompts the user to give a mapping and publishes the result to another
global variable. And a WriteThread which reads from the last global
variable and, when the conversion is complete, shows the results.