Write result to file
First of all, the WriteThread can now also output the result to a file
on the harddisk. Secondly the ConvertThread has been split into a
ConvertThread and a PromptThread so that conversion can continue while
the user is prompted for a mapping.

Also added a second example input file with 524288 lines of data.