Multiple input files and persere input order
Update the tool such that it can now process multiple files at once
(using the same mapping or all files) and also perserve the order of the
input file. It is also possible to merge the results of the multiple
input files into a single file. It is now also possible to export the
mapping generated.

Also added a third example input file with 1048576 lines of data.